swipe/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
Pillow==8.3.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Django==3.2.5
djangorestframework==3.12.4
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
packaging==21.0
Pillow==8.3.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
Pillow==8.3.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
Pillow==8.3.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
gunicorn==20.1.0